Shrunk instruction size, less checks needed

+++ src/instruction.h
@@ -58,8 +58,8 @@ enum bf_opcode {
  */
 struct __attribute__((aligned)) bf_instruction {
     enum bf_opcode opcode;
-    uint32_t argument;
-    uint32_t offset;
+    uint16_t argument;
+    uint16_t offset;
 };
 
 #endif
